// src/interfaces/seats.resp.ts
// Generated by https://quicktype.io

export interface SeatsResp {
    message: string;
    data: Seat[];
}

export interface Seat {
    id: string;
    transaccion: string;
    fecha: string;
    nombre: string;
    zona: string;
    seccion: string;
    area: string;
    fila: number;
    columna: number;
    importe: number;
    descuento: number;
    promocion: string;
    email: string;
    cargo: number;
}